backend/requirements.txt
# requirements.txt - Updated Backend Dependencies
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
werkzeug==3.0.1

# Machine Learning and NLP
torch==2.4.1  # Stable for Windows, compatible with transformers
transformers==4.46.0
pytorch-lightning==2.1.3
datasets==3.1.0
rouge-score==0.1.2
spacy==3.7.2
langdetect==1.0.9

# Data Processing
pandas==2.1.4
numpy==1.26.4  # Compatible with pandas 2.1.4 and Python 3.12
PyPDF2==3.0.1
python-docx==1.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Security and Validation
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-flask==1.3.0

# Development
python-decouple==3.8